FlowField Pathfinding P4 (Subsector Neighbors)
Implemented subsector neighbor detection and storing for use in later pathfinding (subsectors will now find/know which subsectors they're connected to). This calculation is also threaded per sector when a sector is modified.
Added debug drawing for lines between connected subsectors.
Removed subsector numbering to instead have Nodes know their subsector and be able to compare to that instead a numbers. This is also useful as other sectors need to quickly find the owning subsector of an adjacent node.
Sector mesh generation now only applies in editor and will only be calculated whenever debug drawing on A* component is enabled. (This also sped up terrain updates as calculating the mesh took significant time).