Goal (Targeting V1)
Can now set a goal to be a target unit or structure.

If the target can move, the goal will check to recalculate based on the speed of the target and the size of world grid nodes. (gridnode radius / target speed). This way it will only attempt to recalculate if the target has had enough time to move to a new world node. It will also only recalculate if it has moved from the last location at least a distance of grid node diameter.

Added GoalManager to manage goals and be able to call coroutines for the delayed checking of moved targets when necessary.

UnitSelection now casts a ray to check if there's a unit or structure under the mouse in each dimension and will set the goal to that target if one is found.

Fixed static value for pathing directions mismatched.

Began Layers static class to be able to contain constants for use elsewhere (such as layermasks and dimension stuff.

Handled warning for all IntPtr == null always returning true. Now checking for zero properly.

Fixed potential issue in Grid.cpp using 2 << 8 instead of 1 << 8 which would potentially result in checking the wrong dimension or errors.