Structure Goal & Beginning UI
Added ability for Goals to target a structure and still have units path to them instead of not moving.

Added UI for:
 - Resources (not hooked up)
 - Minimap (not hooked up, no logic at all)
 - Unit Selection (mostly functional)
One of each is generated per-player.
Unit Selection is scrollable with 3 columns. Icons are generated with an image from the unit itself. Icon colors change depending on remaining health.